refactor gather_nd ref funcs
