+++ python/tvm/relay/op/_transform.py
@@ -1079,14 +1079,12 @@ def unique_shape_func(attrs, inputs, _):
 @script
 def _gather_nd_shape(data_shape, indices_shape, batch_dims, gather_dim):
     ndim = data_shape.shape[0]
-    mdim = gather_dim
     # using mdim = indices_shape[0] wouldn't work because a rank cannot
     # depend on a runtime shape dimension of indices tensor, even if the
     # dimension is always a known, fixed value. As a workaround, we assume that
     # the fixed gather dimension (the size of an indexing tuple) is recorded
     # in `gather_nd` op attribute.
-    err_msg = "The recorded gather dimension and the actual dimension are different"
-    assert mdim == indices_shape[0], err_msg
+    mdim = gather_dim
     kdim = indices_shape.shape[0] - 1
     out_shape = output_tensor((kdim + ndim - (mdim + batch_dims),), "int64")
     for i in range(1, kdim + 1):
@@ -1101,7 +1099,7 @@ def gather_nd_shape_func(attrs, inputs, _):
     """
     Shape func for ghater_nd operator.
     """
-    batch_dims = get_const_int(attrs.batch_dimss)
+    batch_dims = get_const_int(attrs.batch_dims)
     gather_dim = get_const_int(attrs.gather_dim)
     assert gather_dim > 0, "gather_dim needs to be specified for dynamic gather_nd"
     return [_gather_nd_shape(inputs[0], inputs[1], convert(batch_dims), convert(gather_dim))]
